% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch}
\alias{catch}
\title{catch
This is a data set containing RACEBASE data extract from the GOA and AI}
\format{A data frame with 315048 rows and 10 variables:
\describe{
  \item{HAULJOIN}{unique haul identifier from RACEBASE}
  \item{SPECIES_CODE}{numeric species code}
  \item{WEIGHT}{weight of catch in survey haul}
  \item{NUMBER_FISH}{number of fish in the catch}
  \item{VESSEL}{numeric survey vessel identifier}
  \item{HAUL}{haul number}
  ...
}}
\source{
{RACEBASE}
}
\usage{
catch
}
\description{
catch
This is a data set containing RACEBASE data extract from the GOA and AI
}
\keyword{datasets}
