% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Vast_Index.R
\name{get_VAST_index}
\alias{get_VAST_index}
\title{Get VAST model-based survey index and uncertainty}
\usage{
get_VAST_index(TmbData, Sdreport = Opt[["SD"]], bias.correct,
  Data_Geostat = NULL)
}
\arguments{
\item{TmbData}{Input data object for VAST model}

\item{Sdreport}{Output object with parameter and uncertainty (delta-method) estimates from vAST}

\item{bias.correct}{Boolean for whether bias corrected results should be returned}

\item{Data_Geostat}{}
}
\value{
A table of index values, by year, with associated standard deviation
}
\description{
\code{get_VAST_index} returns table of index values and errors.
}
