% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetPosition.R
\name{NetPosition}
\alias{NetPosition}
\title{Net position}
\usage{
NetPosition(START_LAT, END_LAT, START_LON, END_LON, WIRE_OUT, BOTTOM_DEPTH)
}
\arguments{
\item{START_LAT}{start latitude of a survey tow}

\item{END_LAT}{end latitude of a survey tow}

\item{START_LON}{start longitude of a survey tow}

\item{END_LON}{end longitude of a survey tow}

\item{WIRE_OUT}{wire out during a survey tow}

\item{BOTTOM_DEPTH}{mean bottom depth fo a survey tow}
}
\description{
This function calculates the start and end net positions on the seafloor of a survey
tow using the wire out, depth and start and end latitude and longitude. Uses geometry
and direction of the tow to estimate the offset behind the vessel.
}
\examples{
NetPosition()
}
\keyword{net}
\keyword{positions}
